requirements.txt
faiss==1.5.3
numpy==2.4.4
protobuf==7.34.1
python-dotenv==1.2.2
Requests==2.33.1
sentence_transformers==5.3.0
SQLAlchemy==2.0.25
google-genai==1.46.0